add new folder structure add IsSelected context add ResultsMap refactor for chicago API
