+++ exhibition-curation-fe/src/components/Modal.jsx
