+++ exhibition-curation-fe/src/components/Title.jsx
